+++ src/Page/IMAGE/ImageList.tsx
@@ -6,6 +6,7 @@ import {
 import { Link, useNavigate } from "react-router-dom";
 import { MdOutlineCreateNewFolder, MdDelete, MdEdit } from "react-icons/md";
 import { toast } from "react-toastify";
+import Swal from "sweetalert2";
 
 interface ImageItem {
   _id: string;
@@ -55,8 +56,23 @@ const ImageList: React.FC = () => {
   // Delete Handler
   const handleDelete = async (id: string) => {
     try {
-      await deleteImage(id).unwrap();
-      toast.success("Image deleted successfully!");
+      const result = await Swal.fire({
+        title: "Are you sure?",
+        text: "You won't be able to revert this!",
+        icon: "warning",
+        showCancelButton: true,
+        confirmButtonColor: "#3085d6",
+        cancelButtonColor: "#d33",
+        confirmButtonText: "Yes, delete it!",
+      });
+      if (result.isConfirmed) {
+        await deleteImage(id).unwrap();
+        Swal.fire({
+          title: "Deleted!",
+          text: "Your image has been deleted.",
+          icon: "success",
+        });
+      }
     } catch (err) {
       console.error(err);
       toast.error("Failed to delete image");
@@ -68,17 +84,17 @@ const ImageList: React.FC = () => {
       <div className="max-w-6xl mx-auto">
         {/* Header */}
         <div className="bg-linear-to-r mb-5 from-indigo-500 rounded-t-3xl via-purple-500 to-pink-500 p-4 flex justify-between items-center">
-            <h2 className="text-2xl md:text-3xl font-bold text-white tracking-wide drop-shadow-md">
+          <h2 className="text-2xl md:text-3xl font-bold text-white tracking-wide drop-shadow-md">
             🖼️ Image Dashboard
-            </h2>
+          </h2>
 
-            <Link to="/imageType">
-              <button className="px-5 py-2 flex items-center gap-2  bg-white text-indigo-600 font-semibold rounded-md hover:bg-indigo-50 transition-all shadow-md hover:shadow-lg">
-                <MdOutlineCreateNewFolder className="text-lg" />
-                Create Image
-              </button>
-            </Link>
-          </div>
+          <Link to="/imageType">
+            <button className="px-5 py-2 flex items-center gap-2  bg-white text-indigo-600 font-semibold rounded-md hover:bg-indigo-50 transition-all shadow-md hover:shadow-lg">
+              <MdOutlineCreateNewFolder className="text-lg" />
+              Create Image
+            </button>
+          </Link>
+        </div>
 
         {/* Dashboard Cards */}
         <div className="grid grid-cols-1 sm:grid-cols-2 gap-6 mb-10">
@@ -117,7 +133,7 @@ const ImageList: React.FC = () => {
                 className="bg-white shadow rounded-lg overflow-hidden border border-gray-200 hover:scale-105 transition-transform duration-200 relative cursor-pointer"
               >
                 <img
-                  src={img.image}
+                  src={`http://10.10.7.77:8002${img.image}`}
                   alt={img.title}
                   className="w-full h-48 object-cover"
                   onClick={() => navigate(`/image/${img._id}`)}
